+++ Restaurant/Restaurant/Classes/ApiClasses/Bank.cs
@@ -48,7 +48,7 @@ namespace Restaurant.Classes
             return items;
         }
         
-        public async Task<int> save(Bank item)
+        public async Task<long> save(Bank item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "Banks/Save";
@@ -56,7 +56,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long itemId, long userId, Boolean final)
+        public async Task<long> delete(long itemId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", itemId.ToString());
