+++ Restaurant/Restaurant/Classes/ApiClasses/Branch.cs
@@ -221,7 +221,7 @@ namespace Restaurant.Classes
         //    }
         //    return items;
         //}
-        public async Task<int> save(Branch item)
+        public async Task<long> save(Branch item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "Branches/Save";
@@ -230,7 +230,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long branchId, long userId, Boolean final)
+        public async Task<long> delete(long branchId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", branchId.ToString());
