+++ Restaurant/Restaurant/Classes/ApiClasses/Card.cs
@@ -46,7 +46,7 @@ namespace Restaurant.Classes
             return items;
         }
               
-        public async Task<int> save(Card item)
+        public async Task<long> save(Card item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "cards/Save";
@@ -54,7 +54,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long cardId, long userId, Boolean final)
+        public async Task<long> delete(long cardId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", cardId.ToString());
