+++ Restaurant/Restaurant/Classes/ApiClasses/Coupon.cs
@@ -130,7 +130,7 @@ namespace Restaurant.Classes
             return item;
         }
       
-        public async Task<int> save(Coupon item)
+        public async Task<long> save(Coupon item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "coupons/Save";
@@ -138,7 +138,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long itemId, long userId, Boolean final)
+        public async Task<long> delete(long itemId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", itemId.ToString());
