+++ Restaurant/Restaurant/Classes/ApiClasses/DishIngredients.cs
@@ -45,7 +45,7 @@ namespace Restaurant.Classes
             }
             return items;
         }
-        public async Task<int> save(DishIngredients item)
+        public async Task<long> save(DishIngredients item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "dishIngredients/Save";
@@ -53,7 +53,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long itemId, long userId, Boolean final)
+        public async Task<long> delete(long itemId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", itemId.ToString());
