+++ Restaurant/Restaurant/Classes/ApiClasses/Inventory.cs
@@ -94,7 +94,7 @@ namespace Restaurant.Classes
             }
             return IsManipulated;
         }
-        public async Task<int> save(Inventory item)
+        public async Task<long> save(Inventory item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "Inventory/Save";
@@ -103,7 +103,7 @@ namespace Restaurant.Classes
            return await APIResult.post(method, parameters);
         }
        
-        public async Task<int> delete(long itemId, long userId, Boolean final)
+        public async Task<long> delete(long itemId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", itemId.ToString());
