+++ Restaurant/Restaurant/Classes/ApiClasses/InvoicesClass.cs
@@ -76,7 +76,7 @@ namespace Restaurant.Classes
             }
             return item;
         }
-        public async Task<int> save(InvoicesClass item)
+        public async Task<long> save(InvoicesClass item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "invoicesClass/Save";
@@ -84,7 +84,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long itemId, long userId, Boolean final)
+        public async Task<long> delete(long itemId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", itemId.ToString());
