+++ Restaurant/Restaurant/Classes/ApiClasses/Offer.cs
@@ -65,7 +65,7 @@ namespace Restaurant.Classes
             }
             return item;
         }
-        public async Task<int> save(Offer item)
+        public async Task<long> save(Offer item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "Offers/Save";
@@ -73,7 +73,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long itemId, long userId, Boolean final)
+        public async Task<long> delete(long itemId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", itemId.ToString());
