+++ Restaurant/Restaurant/Classes/ApiClasses/Pos.cs
@@ -81,7 +81,7 @@ namespace Restaurant.Classes
             }
             return item;
         }
-        public async Task<int> save(Pos item)
+        public async Task<long> save(Pos item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "Pos/Save";
@@ -89,7 +89,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long itemId, long userId, Boolean final)
+        public async Task<long> delete(long itemId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", itemId.ToString());
