+++ Restaurant/Restaurant/Classes/ApiClasses/Section.cs
@@ -83,7 +83,7 @@ namespace Restaurant.Classes
             }
             return items;
         }
-        public async Task<int> save(Section item)
+        public async Task<long> save(Section item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "Sections/Save";
@@ -91,7 +91,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long sectionId, long userId, Boolean final)
+        public async Task<long> delete(long sectionId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", sectionId.ToString());
