+++ Restaurant/Restaurant/Classes/ApiClasses/SubscriptionFees.cs
@@ -45,7 +45,7 @@ namespace Restaurant.Classes
         }
 
       
-        public async Task<int> save(SubscriptionFees item)
+        public async Task<long> save(SubscriptionFees item)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             string method = "subscriptionFees/Save";
@@ -53,7 +53,7 @@ namespace Restaurant.Classes
             parameters.Add("itemObject", myContent);
            return await APIResult.post(method, parameters);
         }
-        public async Task<int> delete(long itemId, long userId, Boolean final)
+        public async Task<long> delete(long itemId, long userId, Boolean final)
         {
             Dictionary<string, string> parameters = new Dictionary<string, string>();
             parameters.Add("itemId", itemId.ToString());
